move functions to crossover file
